+++ lib/matrix.dart
@@ -6,9 +6,9 @@ import 'package:ml_linalg/src/matrix/float32x4/float32x4_matrix.dart';
 import 'package:ml_linalg/vector.dart';
 
 /// An algebraic matrix
-abstract class MLMatrix {
+abstract class Matrix {
   /// Creates a matrix from a two dimensional list
-  factory MLMatrix.from(List<List<double>> source, {Type dtype = Float32x4}) {
+  factory Matrix.from(List<List<double>> source, {Type dtype = Float32x4}) {
     switch (dtype) {
       case Float32x4:
         return Float32x4Matrix.from(source);
@@ -18,7 +18,7 @@ abstract class MLMatrix {
   }
 
   /// Creates a matrix with predefined row vectors
-  factory MLMatrix.rows(List<MLVector> source, {Type dtype = Float32x4}) {
+  factory Matrix.rows(List<Vector> source, {Type dtype = Float32x4}) {
     switch (dtype) {
       case Float32x4:
         return Float32x4Matrix.rows(source);
@@ -28,7 +28,7 @@ abstract class MLMatrix {
   }
 
   /// Creates a matrix with predefined column vectors
-  factory MLMatrix.columns(List<MLVector> source, {Type dtype = Float32x4}) {
+  factory Matrix.columns(List<Vector> source, {Type dtype = Float32x4}) {
     switch (dtype) {
       case Float32x4:
         return Float32x4Matrix.columns(source);
@@ -36,6 +36,8 @@ abstract class MLMatrix {
         throw UnimplementedError();
     }
   }
+
+  Type get dtype;
 
   /// Returns a number of matrix row
   int get rowsNum;
@@ -47,61 +49,61 @@ abstract class MLMatrix {
   Iterable<double> operator [](int index);
 
   /// Performs sum of the matrix and a matrix/ a vector/ a scalar/ whatever
-  MLMatrix operator +(Object value);
+  Matrix operator +(Object value);
 
   /// Performs subtraction of the matrix and a matrix/ a vector/ a scalar/
   /// whatever
-  MLMatrix operator -(Object value);
+  Matrix operator -(Object value);
 
   /// Performs multiplication of the matrix and a matrix/ a vector/ a scalar/
   /// whatever
-  MLMatrix operator *(Object value);
+  Matrix operator *(Object value);
 
   /// Performs division of the matrix by a matrix/ a vector/ a scalar/
   /// whatever
-  MLMatrix operator /(Object value);
+  Matrix operator /(Object value);
 
   /// Performs transposition of the matrix
-  MLMatrix transpose();
+  Matrix transpose();
 
   /// Cuts out a part of the matrix bounded by [rows] and [columns] range
-  MLMatrix submatrix({Range rows, Range columns});
+  Matrix submatrix({Range rows, Range columns});
 
   /// Creates a new matrix, consisted of different segments of the matrix
-  MLMatrix pick({Iterable<Range> rowRanges, Iterable<Range> columnRanges});
+  Matrix pick({Iterable<Range> rowRanges, Iterable<Range> columnRanges});
 
   /// Returns a column of the matrix, resided on [index]
-  MLVector getColumn(int index, {bool tryCache = true, bool mutable = false});
+  Vector getColumn(int index, {bool tryCache = true, bool mutable = false});
 
   /// Returns a row of the matrix, resided on [index]
-  MLVector getRow(int index, {bool tryCache = true, bool mutable = false});
+  Vector getRow(int index, {bool tryCache = true, bool mutable = false});
 
   /// Reduces all the matrix columns to only column, using [combiner] function
-  MLVector reduceColumns(MLVector combiner(MLVector combine, MLVector vector),
-      {MLVector initValue});
+  Vector reduceColumns(Vector combiner(Vector combine, Vector vector),
+      {Vector initValue});
 
   /// Reduces all the matrix rows to only row, using [combiner] function
-  MLVector reduceRows(MLVector combiner(MLVector combine, MLVector vector),
-      {MLVector initValue});
+  Vector reduceRows(Vector combiner(Vector combine, Vector vector),
+      {Vector initValue});
 
   /// Performs column-wise mapping of this matrix to a new one via passed
   /// [mapper] function
-  MLMatrix mapColumns(MLVector mapper(MLVector column));
+  Matrix mapColumns(Vector mapper(Vector column));
 
   /// Performs row-wise mapping of this matrix to a new one via passed
   /// [mapper] function
-  MLMatrix mapRows(MLVector mapper(MLVector row));
+  Matrix mapRows(Vector mapper(Vector row));
 
   /// Creates a new matrix, efficiently iterating through all the matrix
   /// elements (several floating point elements in a time) and applying the
   /// [mapper] function
-  MLMatrix fastMap<E>(E mapper(E columnElement));
+  Matrix fastMap<E>(E mapper(E columnElement));
 
   /// Tries to convert the matrix to a vector.
   ///
   /// It fails, if the matrix's both numbers of columns and rows are greater
   /// than `1`
-  MLVector toVector({bool mutable = false});
+  Vector toVector({bool mutable = false});
 
   /// Returns max value of the matrix
   double max();
@@ -118,4 +120,7 @@ abstract class MLMatrix {
   /// [columnValues] - values, that are going to be placed one by one in the
   /// target column
   void setColumn(int columnNum, Iterable<double> columnValues);
+
+  /// Extracts non-repeated matrix rows and pack them into matrix
+  Matrix uniqueRows();
 }