+++ lib/src/matrix/float32x4/float32x4_matrix_factory_mixin.dart
@@ -1,23 +1,23 @@
 import 'package:ml_linalg/matrix.dart';
 import 'package:ml_linalg/src/matrix/float32x4/float32x4_matrix.dart';
-import 'package:ml_linalg/src/matrix/ml_matrix_factory.dart';
+import 'package:ml_linalg/src/matrix/matrix_factory.dart';
 import 'package:ml_linalg/vector.dart';
 
-abstract class Float32x4MatrixFactoryMixin implements MLMatrixFactory {
+abstract class Float32x4MatrixFactoryMixin implements MatrixFactory {
   @override
-  MLMatrix createMatrixFrom(Iterable<Iterable<double>> source) =>
+  Matrix createMatrixFrom(Iterable<Iterable<double>> source) =>
       Float32x4Matrix.from(source);
 
   @override
-  MLMatrix createMatrixFromFlattened(
+  Matrix createMatrixFromFlattened(
           Iterable<double> source, int rowsNum, int columnsNum) =>
       Float32x4Matrix.flattened(source, rowsNum, columnsNum);
 
   @override
-  MLMatrix createMatrixFromRows(Iterable<MLVector> source) =>
+  Matrix createMatrixFromRows(Iterable<Vector> source) =>
       Float32x4Matrix.rows(source);
 
   @override
-  MLMatrix createMatrixFromColumns(Iterable<MLVector> source) =>
+  Matrix createMatrixFromColumns(Iterable<Vector> source) =>
       Float32x4Matrix.columns(source);
 }