+++ lib/src/matrix/matrix_data_store.dart
@@ -2,9 +2,9 @@ import 'dart:typed_data';
 
 import 'package:ml_linalg/vector.dart';
 
-abstract class MLMatrixDataStore {
-  List<MLVector> get columnsCache;
-  List<MLVector> get rowsCache;
+abstract class MatrixDataStore {
+  List<Vector> get columnsCache;
+  List<Vector> get rowsCache;
   ByteData get data;
   void updateByteData(int elementNum, double value);
 }