+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 6.0.0
+- prefix `ML` removed from entities' names
+- Float32x4Vector: equality operator override added
+- Matrix: `uniqueRows` method added
+
 ## 5.5.1
 - MLMatrix: fixed bug in `setColumn` method when rows cache was not cleared